backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
astroid==3.3.9
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
dill==0.3.9
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-ninja==1.3.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
fido2==1.2.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
multidict==6.2.0
oauthlib==3.2.2
packaging==24.2
platformdirs==4.3.7
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pylint==3.3.6
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-http-client==3.3.7
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tomlkit==0.13.2
twilio==9.5.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3


# Django==3.2.13
# djangorestframework==3.13.1
# PyJWT==2.4.0
# django-cors-headers==3.13.0
# gunicorn==20.1.0
# django-environ==0.8.1
